+++ app.js
@@ -46,7 +46,6 @@ helpersInitialized.then(function (o) {
     redirectIfApiDomain,
     redirectIfHasZidButNoConversationId,
     redirectIfNotHttps,
-    resolve_pidThing,
     signInJoin,
     timeout,
     winston,
@@ -205,6 +204,7 @@ helpersInitialized.then(function (o) {
     need,
     needCookie,
     needHeader,
+    resolve_pidThing,
     want,
     wantCookie,
     wantHeader,
